requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg2-binary==2.9.9
paho-mqtt==2.1.0
python-dotenv==1.0.1
openai==1.42.0
python-dateutil==2.9.0.post0